regenerate vmlinux.rs for 6.12.6
This is to accomodate the structs changing between kernel versions. Aya
does not have support for emitting code that is CO:RE( Compile Once :
Run Everywhere ) yet. It can **run** CO:RE code.

A way around this is to write c helpers that utilize CO:RE and linking
against them.